+++ src/plugins/extendConfig.ts
@@ -3,6 +3,7 @@ import * as path from "node:path"
 import typescript from "@rollup/plugin-typescript"
 import { defuFn } from "defu"
 import renameNodeModules from "rollup-plugin-rename-node-modules"
+import ts from "typescript"
 import type { Plugin, UserConfig } from "vite"
 import { defineConfig } from "vite"
 
@@ -60,6 +61,9 @@ export const extendConfigPlugin = (options: Options): Plugin => {
 							declaration: options.dts,
 							outDir: userConfig.build?.outDir || "dist",
 							include: [path.posix.join(options.srcDir, "**/*")],
+							transformers: {
+								afterDeclarations: [addImportExportExtensionsTransformer],
+							},
 						}) as Plugin,
 						// Ensure bundled dependencies are published to npm.
 						// `npm pack` ignores directories named `node_modules`.
@@ -82,5 +86,63 @@ export const extendConfigPlugin = (options: Options): Plugin => {
 
 			return defuFn(userConfig, getDefaultConfig(userConfig))
 		},
+	}
+}
+
+/**
+ * Adds a `.js` extension to all import and export declarations that do not
+ * contain an extension.
+ */
+function addImportExportExtensionsTransformer(
+	context: ts.TransformationContext,
+) {
+	return (source: ts.Bundle | ts.SourceFile) => {
+		function visitor(node: ts.Node) {
+			function shouldAddExtension(moduleSpecifier: ts.StringLiteral) {
+				const text = moduleSpecifier.text
+
+				return text.startsWith(".") && !/\.[a-z0-9]+$/i.test(text)
+			}
+
+			function addExtension(moduleSpecifier: ts.StringLiteral) {
+				return ts.factory.createStringLiteral(moduleSpecifier.text + ".js")
+			}
+
+			if (ts.isImportDeclaration(node)) {
+				if (
+					ts.isStringLiteral(node.moduleSpecifier) &&
+					shouldAddExtension(node.moduleSpecifier)
+				) {
+					return ts.factory.updateImportDeclaration(
+						node,
+						node.modifiers,
+						node.importClause,
+						addExtension(node.moduleSpecifier),
+						node.attributes,
+					)
+				}
+			}
+
+			if (ts.isExportDeclaration(node)) {
+				if (
+					node.moduleSpecifier &&
+					ts.isStringLiteral(node.moduleSpecifier) &&
+					shouldAddExtension(node.moduleSpecifier)
+				) {
+					return ts.factory.updateExportDeclaration(
+						node,
+						node.modifiers,
+						node.isTypeOnly,
+						node.exportClause,
+						addExtension(node.moduleSpecifier),
+						node.attributes,
+					)
+				}
+			}
+
+			return ts.visitEachChild(node, visitor, context)
+		}
+
+		return ts.visitEachChild(source, visitor, context)
 	}
 }